+++ packages/form/lib/src/extensions/form_state_field_extension.dart
@@ -27,6 +27,7 @@ import 'package:bond_form/bond_form.dart';
 /// - [asyncDropDownField] Retrieves the state of an async dropdown field.
 /// - [asyncDropDownItems] Retrieves a list of async dropdown items.
 /// - [hiddenField] Retrieves the state of a hidden field.
+/// - [asyncHiddenField] Retrieves the state of an async hidden field.
 /// - [fileField] Retrieves the state of a file field.
 extension FieldBondFormState on BaseBondFormState {
   /// Retrieves the [TextFieldState] for a given text field.
@@ -182,6 +183,15 @@ extension FieldBondFormState on BaseBondFormState {
   /// - Throws: [ArgumentError] if the field doesn't exist.
   HiddenFieldState<T> hiddenField<T>(String fieldName) {
     return get<HiddenFieldState<T>, T>(fieldName);
+  }
+
+  /// Retrieves the async hidden field state for a specified [fieldName].
+  ///
+  /// - Parameter [fieldName] The name of the async hidden field to fetch.
+  /// - Returns: The async hidden field state.
+  /// - Throws: [ArgumentError] if the field doesn't exist.
+  AsyncHiddenFieldState<T> asyncHiddenField<T>(String fieldName) {
+    return get<AsyncHiddenFieldState<T>, T>(fieldName);
   }
 
   /// Retrieves the state of the file field for a specified [fieldName].