+++ packages/form/lib/src/form_fields.dart
@@ -1,6 +1,7 @@
 library form_fields;
 
 export 'form_fields/fields/async_drop_down_field_state.dart';
+export 'form_fields/fields/async_hidden_field_state.dart';
 export 'form_fields/fields/async_radio_group_field_state.dart';
 export 'form_fields/fields/checkbox_field_state.dart';
 export 'form_fields/fields/checkbox_group_field_state.dart';